new idea
i am now making a neural network from scratch that is trained on social media posts in order to see if the post is toxic or non toxic. i will use this in a discord server to 'kick' people who are saying some bad stuff. since it is a neural network, it will have some advantages such as being able to tell the difference between cursing at someone and just saying it in another context (ex: this is so ****ing awesome does not equal 'I ****ing hate you) some of the coments on the dataset are pretty bad so i may have to lower the threshhold so that it picks up on more stuff. i have made the tokenizer, and am working on formatting the data. i still need to build out the vocabulary and format the data i little better, but it is going pretty well so far. The goal it to use minimal ai to code, but maybe use it to learn some js syntax or common practices in creating a neural network from scratch. So far i have learned a lot about JS, data, csv files, shuffling algorythms, and vector embeddings. More to come!